+++ src/components/morinfocard/moreinfocard.tsx
@@ -8,14 +8,14 @@ import {
   MDBCol,
   MDBRipple,
   MDBRow,
+  MDBTextArea,
+  MDBInput,
 } from 'mdb-react-ui-kit';
-import { useEffect, useState } from 'react';
-import { propTypes } from 'react-bootstrap/esm/Image';
+import { useEffect, useState, SyntheticEvent } from 'react';
 import { useParams } from 'react-router-dom';
 import { Recipe } from '../../models/recipe';
 import { User } from '../../models/user';
 import Reviews from '../reviews/reviews';
-import Update from '../update/update';
 
 interface IMoreInfoProps {
   currentUser: User | undefined;
@@ -25,19 +25,46 @@ interface IMoreInfoProps {
 }
 
 function MoreInfoCard(props: IMoreInfoProps) {
+
   const [authUser, setAuthUser] = useState<User>();
   const [recipe, setRecipe] = useState<any>({});
-  const [editing, setEditing] = useState(false); //needed for conditional rendering of update component
+  const [editing, setEditing] = useState(false);
+
   let { id } = useParams();
 
   function handleEditClick(){
     setEditing(!editing);
-  }
-
-   function handleEditSubmit(){
     getRecipe();
+      //@DOCS: THIS WAY, OUR CHANGES ARE ERASED ON HITTING 'CANCEL', BUT PERSISTED IF 
+      //WE HIT SAVE.
   }
 
+
+  function handleChange(e: SyntheticEvent){
+    setRecipe({
+        ...recipe,
+        [(e.target as HTMLInputElement).name]: (e.target as HTMLInputElement).value
+    })
+}
+
+  function handleSubmit(){
+    fetch(`http://localhost:8080/recipes/${id}`, { 
+    method: 'PATCH',
+    headers: { 
+        'Content-type': 'application/json',
+        'Access-Control-Allow-Origin': "*"
+      },
+    body: JSON.stringify( {...recipe}  ),
+}).then((res) => {
+    const newData = res.json();
+    setRecipe(recipe); //@DOCS: THIS IS THE MAGIC BIT. 
+                      // DO NOT CHANGE ME OR BAD UGLY THINGS
+                     //  WILL HAPPEN!
+})
+setEditing(!editing);
+}
+
+  
   async function getRecipe() {
     const res = await fetch(`http://localhost:8080/recipes/${id}`, {
       method: 'GET',
@@ -50,9 +77,13 @@ function MoreInfoCard(props: IMoreInfoProps) {
     setRecipe(Object.assign(data));
   }
 
+  
+
   useEffect(() => {
     getRecipe();
   }, []);
+
+ 
 
   return recipe ? (
     <MDBCard>
@@ -81,17 +112,45 @@ function MoreInfoCard(props: IMoreInfoProps) {
           </MDBCardTitle>
           <MDBCardText>{recipe.instructions}</MDBCardText>
           <MDBBtn style={{width: '12.5%'}} onClick={handleEditClick}> {editing ? 'CANCEL' : 'EDIT'}</MDBBtn>
-          {editing ? <Update 
-        id={id}
-        author={recipe.author}
-        recipe_name={recipe.recipe_name}
-        instructions={recipe.instructions}
-        category={recipe.category}
-        editing={editing}
-        recipe={recipe}
-        setRecipe={setRecipe}
-        handleSubmitClick={handleEditSubmit}
-        /> : <></>}
+          <div style={{ margin: 'auto', textAlign: 'center', width: '80%' }}>
+             {editing &&
+             <div className='edit-true'> 
+             <form onSubmit={handleSubmit}>
+             <label>Recipe name</label>
+             <MDBInput 
+             value={recipe.recipe_name}
+             id="recipe-name"
+             name="recipe_name"
+             className="small-top-margin"
+             onChange={handleChange}
+             />
+ 
+             <label>Instructions</label>
+             <MDBTextArea 
+             value={recipe.instructions}
+             id="instructions"
+             name="instructions"
+             className="small-top-margin"
+             rows={9}
+             onChange={handleChange}
+             />
+ 
+             <label>Category</label>
+             <input 
+             value={recipe.category}
+             id="category"
+             name="category"
+             onChange={handleChange}
+             />
+ 
+             </form>
+ 
+             <MDBBtn className="small-top-margin" onClick={handleSubmit}>SAVE your CHANGES!</MDBBtn>
+             <br></br>
+             </div> 
+             }
+            
+        </div>
         </MDBCol>
         <MDBCol>
           <h2>Reviews</h2>
