--- src/components/update/update.css
@@ -1,7 +1,0 @@
-.small-top-margin {
-    margin-top: 1rem;
-}
-
-.small-bottom-margin {
-    margin-bottom: 1rem;
-}
